Fix custom DNS servers
Previously the DNS servers were set in the cloud-init only, which worked
for the initial server bootstrap, but later these DNS servers get
replaced by NetworkManager.

This patch fixes this behaviour by saving the custom DNS servers in
NetworkManager which has the added benefit of enabling user to change it
after deployment.

Signed-off-by: Marek Pikuła <marek@serenitycode.dev>